+++ BxInputController/Sources/Controller/BxInputController.swift
@@ -27,7 +27,7 @@ open class BxInputController : UIViewController
         didSet { updateEstimatedContent() }
     }
     // This changed automatical, when changed isEstimatedContent value.
-    var tableDelegate: BaseTableDelegate? = nil
+    var tableDelegate: BxInputBaseTableDelegate? = nil
     
     internal(set) public var datePicker: UIDatePicker = UIDatePicker()
     internal(set) public var variantsPicker: UIPickerView = UIPickerView()
@@ -185,13 +185,13 @@ open class BxInputController : UIViewController
     {
         tableView.dataSource = self
         if isEstimatedContent {
-            tableDelegate = EstimatedSizeTableDelegate(parent: self)
+            tableDelegate = BxInputEstimatedSizeTableDelegate(parent: self)
             
             tableView.rowHeight = UITableViewAutomaticDimension
             tableView.sectionHeaderHeight = UITableViewAutomaticDimension
             tableView.sectionFooterHeight = UITableViewAutomaticDimension
         } else {
-            tableDelegate = StaticSizeTableDelegate(parent: self)
+            tableDelegate = BxInputStaticSizeTableDelegate(parent: self)
             
             tableView.rowHeight = 60
             tableView.sectionHeaderHeight = 20